pyproject.toml
[project]
name = "octofin"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "djp (>=0.3.1,<0.4.0)",
    "djangorestframework (>=3.16.0,<4.0.0)"
]

[tool.poetry]
packages = [{include = "octofin", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
